fazendo ajustes

+++ backend/talkpdf/views.py
@@ -251,16 +251,32 @@ class PDFChatView(APIView):
         # Se novos PDFs forem fornecidos, processe-os
         if pdf_files:
             try:
-                if os.path.exists(chat.path):
-                    shutil.rmtree(chat.path, ignore_errors=True)
+                # Gerar um nome único para o diretório temporário
+                temp_folder_name = f"temp_{uuid.uuid4()}"
+                temp_folder_path = os.path.join(os.path.dirname(chat.path), temp_folder_name)
 
-                # Extraia os nomes dos arquivos e atualize no banco de dados
+                # Processar os novos PDFs e gerar os novos vetores no diretório temporário
+                folder_path = get_vectorstore_from_files(pdf_files, thread_id, temp_dir=temp_folder_path)
+
+                # Verificar se a criação dos vetores foi bem-sucedida
+                if not os.path.exists(folder_path):
+                    raise Exception("Falha na criação do novo vectorstore.")
+
+                # Extrair os nomes dos arquivos e atualizar no banco de dados
                 file_names = [file.name for file in pdf_files]
                 chat.file_names = file_names
 
-                folder_path = get_vectorstore_from_files(pdf_files, thread_id)
+                # Apagar o diretório antigo após a criação bem-sucedida dos novos vetores
+                if os.path.exists(chat.path):
+                    shutil.rmtree(chat.path, ignore_errors=True)
+
+                # Atualizar o caminho do chat para o novo folder_path
                 chat.path = folder_path
+
             except Exception as e:
+                # Remover o diretório temporário caso exista para evitar acumulação de dados
+                if os.path.exists(temp_folder_path):
+                    shutil.rmtree(temp_folder_path, ignore_errors=True)
                 return Response({'error': f'Erro ao processar os PDFs: {str(e)}'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
         # Salvar as alterações no chat
